src/main/java/ro/gdgs/crawler/web/GreetServlet.java
package ro.gdgs.crawler.web;

import javax.servlet.ServletException;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;
import java.io.PrintWriter;

/**
 * @author Octavian
 * @since 1.0
 */
public class GreetServlet extends HttpServlet {
    @Override
    public void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
        response.setContentType("text/plain");
        response.setCharacterEncoding("UTF-8");

        PrintWriter out = response.getWriter();
        try {
            String name = request.getParameter("username");
            out.println("Salut " + name);
        } finally {
            out.close();
        }
    }
}
